+++ internal/ceres/block_sparse_matrix.h
@@ -113,9 +113,9 @@ class CERES_NO_EXPORT BlockSparseMatrix final : public SparseMatrix {
   void AddTransposeBlockStructure();
 
   // clang-format off
-  int num_rows()         const final { return num_rows_;     }
-  int num_cols()         const final { return num_cols_;     }
-  int num_nonzeros()     const final { return num_nonzeros_; }
+  int64_t num_rows()         const final { return num_rows_;     }
+  int64_t num_cols()         const final { return num_cols_;     }
+  int64_t num_nonzeros()     const final { return num_nonzeros_; }
   const double* values() const final { return values_; }
   double* mutable_values()     final { return values_; }
   // clang-format on
@@ -162,14 +162,14 @@ class CERES_NO_EXPORT BlockSparseMatrix final : public SparseMatrix {
       bool use_page_locked_memory = false);
 
  private:
-  double* AllocateValues(int size);
+  double* AllocateValues(int64_t size);
   void FreeValues(double*& values);
 
   const bool use_page_locked_memory_;
-  int num_rows_;
-  int num_cols_;
-  int num_nonzeros_;
-  int max_num_nonzeros_;
+  int64_t num_rows_;
+  int64_t num_cols_;
+  int64_t num_nonzeros_;
+  int64_t max_num_nonzeros_;
   double* values_;
   std::unique_ptr<CompressedRowBlockStructure> block_structure_;
   std::unique_ptr<CompressedRowBlockStructure> transpose_block_structure_;
