+++ internal/ceres/detect_structure.h
@@ -57,7 +57,7 @@ namespace ceres::internal {
 // the values returned by this function. It is entirely possible that
 // the f_block_size and row_blocks_size is not constant in such rows.
 void CERES_NO_EXPORT DetectStructure(const CompressedRowBlockStructure& bs,
-                                     const int num_eliminate_blocks,
+                                     const int64_t num_eliminate_blocks,
                                      int* row_block_size,
                                      int* e_block_size,
                                      int* f_block_size);
